tests: a payload materialized inside a macro scope, asked whether it survives
A macro expansion's reclaim balances the references a scan of heap contents
cannot explain, and the payload cache's reference is held in Rust, out of that
scan's reach. Both tests are red: the reclaim takes the cache's reference, the
region goes with the expansion's own headers, and the cache goes on naming it.

The second one holds the shape a running program takes. Payloads pack several
to a region, so an expansion opens a fresh one only when the open region passes
its size threshold — which is why a stale payload surfaces after a long run
rather than at any one macro. A fresh heap opens its first payload region on
the first materialization, so it cannot tell that case from this one.

The file is split at its two subjects to stay inside the reading budget: which
payload a header reads and how long its region lives, and what a payload
carries.
